requirements.txt
# API and Server
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai>=1.0.0
tiktoken>=0.5.0

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
black==23.11.0
pytest==7.4.3
pytest-asyncio==0.21.1
